refactor: break up src/vm/scope.rs into focused modules
Split the 481-line monolithic scope.rs into a structured module directory:
- runtime_scope.rs: RuntimeScope struct for individual scope levels
- scope_stack.rs: ScopeStack struct for managing scope hierarchy
- handlers.rs: Instruction handlers for scope management (PushScope, PopScope, etc)
- mod.rs: Module organization with re-exports for backward compatibility

This improves maintainability by separating concerns while keeping the public
API unchanged. All tests pass and the code is properly formatted.
