+++ src/vm/scope/mod.rs
@@ -1,0 +1,10 @@
+pub mod handlers;
+mod runtime_scope;
+mod scope_stack;
+
+pub use handlers::{
+    handle_define_local, handle_load_scoped, handle_pop_scope, handle_push_scope,
+    handle_store_scoped,
+};
+pub use runtime_scope::RuntimeScope;
+pub use scope_stack::ScopeStack;
