+++ src/vm/scope/scope_stack.rs
@@ -1,53 +1,6 @@
+use super::runtime_scope::RuntimeScope;
 use crate::compiler::scope::ScopeType;
 use crate::value::Value;
-use std::collections::HashMap;
-
-// These handlers will be defined below
-// For now, just declare them as they'll be implemented step by step
-
-/// Represents a single scope level at runtime
-#[derive(Debug, Clone)]
-pub struct RuntimeScope {
-    /// Variables defined at this scope level (symbol_id -> Value)
-    pub variables: HashMap<u32, Value>,
-    /// Type of this scope
-    pub scope_type: ScopeType,
-}
-
-impl RuntimeScope {
-    /// Create a new runtime scope
-    pub fn new(scope_type: ScopeType) -> Self {
-        RuntimeScope {
-            variables: HashMap::new(),
-            scope_type,
-        }
-    }
-
-    /// Define a variable in this scope
-    pub fn define(&mut self, sym_id: u32, value: Value) {
-        self.variables.insert(sym_id, value);
-    }
-
-    /// Get a variable from this scope
-    pub fn get(&self, sym_id: u32) -> Option<&Value> {
-        self.variables.get(&sym_id)
-    }
-
-    /// Get a mutable reference to a variable
-    pub fn get_mut(&mut self, sym_id: u32) -> Option<&mut Value> {
-        self.variables.get_mut(&sym_id)
-    }
-
-    /// Set a variable in this scope (returns old value if present)
-    pub fn set(&mut self, sym_id: u32, value: Value) -> Option<Value> {
-        self.variables.insert(sym_id, value)
-    }
-
-    /// Check if a variable is defined in this scope
-    pub fn contains(&self, sym_id: u32) -> bool {
-        self.variables.contains_key(&sym_id)
-    }
-}
 
 /// Manages the runtime scope stack during execution
 pub struct ScopeStack {
@@ -189,119 +142,9 @@ impl Default for ScopeStack {
     }
 }
 
-/// Instruction handlers for scope management
-/// These are called by the VM to manage runtime scopes
-use crate::vm::core::VM;
-
-/// Handle PushScope instruction
-pub fn handle_push_scope(vm: &mut VM, scope_type_byte: u8) -> Result<(), String> {
-    // Convert byte to ScopeType
-    let scope_type = match scope_type_byte {
-        0 => ScopeType::Global,
-        1 => ScopeType::Function,
-        2 => ScopeType::Block,
-        3 => ScopeType::Loop,
-        4 => ScopeType::Let,
-        _ => return Err(format!("Invalid scope type: {}", scope_type_byte)),
-    };
-
-    vm.scope_stack.push(scope_type);
-    Ok(())
-}
-
-/// Handle PopScope instruction
-pub fn handle_pop_scope(vm: &mut VM) -> Result<(), String> {
-    if !vm.scope_stack.pop() {
-        return Err("Cannot pop global scope".to_string());
-    }
-    Ok(())
-}
-
-/// Handle LoadScoped instruction
-pub fn handle_load_scoped(_vm: &mut VM, bytecode: &[u8], ip: &mut usize) -> Result<(), String> {
-    let depth = bytecode[*ip] as usize;
-    *ip += 1;
-    let index = bytecode[*ip] as usize;
-    *ip += 1;
-
-    // This instruction is for future use - currently variables use LoadUpvalue
-    // For now, just treat as a no-op to avoid breaking existing code
-    let _ = depth;
-    let _ = index;
-    Ok(())
-}
-
-/// Handle StoreScoped instruction
-pub fn handle_store_scoped(vm: &mut VM, bytecode: &[u8], ip: &mut usize) -> Result<(), String> {
-    let depth = bytecode[*ip] as usize;
-    *ip += 1;
-    let index = bytecode[*ip] as usize;
-    *ip += 1;
-
-    // Pop value from stack
-    let value = vm.stack.pop().ok_or("Stack underflow")?;
-
-    // Store to scope at the specified depth
-    if !vm.scope_stack.set_at_depth(depth, index as u32, value) {
-        return Err(format!(
-            "Variable not found at depth {} index {}",
-            depth, index
-        ));
-    }
-
-    Ok(())
-}
-
-/// Handle DefineLocal instruction
-pub fn handle_define_local(
-    vm: &mut VM,
-    bytecode: &[u8],
-    ip: &mut usize,
-    constants: &[Value],
-) -> Result<(), String> {
-    // Read symbol index from bytecode
-    let high = bytecode[*ip] as u16;
-    let low = bytecode[*ip + 1] as u16;
-    *ip += 2;
-    let sym_idx = (high << 8) | low;
-
-    // Pop value from stack
-    let value = vm.stack.pop().ok_or("Stack underflow")?;
-
-    // Get the symbol ID from constants
-    let sym_id = if let Value::Symbol(id) = constants[sym_idx as usize] {
-        id.0
-    } else {
-        return Err("Expected symbol in constants".to_string());
-    };
-
-    // Define in current scope
-    vm.scope_stack.define_local(sym_id, value);
-
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_runtime_scope_creation() {
-        let scope = RuntimeScope::new(ScopeType::Global);
-        assert_eq!(scope.scope_type, ScopeType::Global);
-        assert!(scope.variables.is_empty());
-    }
-
-    #[test]
-    fn test_runtime_scope_define_and_get() {
-        let mut scope = RuntimeScope::new(ScopeType::Block);
-        let sym_id = 42;
-        let value = Value::Int(123);
-
-        scope.define(sym_id, value.clone());
-        assert_eq!(scope.get(sym_id), Some(&value));
-        assert!(scope.contains(sym_id));
-    }
 
     #[test]
     fn test_scope_stack_creation() {
